group section and simplified UI
