+++ example/src/Navigation.tsx
@@ -2,7 +2,7 @@ import { createNativeStackNavigator } from '@react-navigation/native-stack'
 
 export type NavigationParamList = {
   launch: undefined
-  test: undefined
+  test: { onlyGroups?: boolean}
   home: undefined
   group: { id: string }
   conversation: { topic: string }
