+++ example/src/TestScreen.tsx
@@ -1,7 +1,9 @@
 import React, { useEffect, useState } from 'react'
 import { View, Text, Button, ScrollView } from 'react-native'
+import { useRoute } from '@react-navigation/native';
 
-import { tests, Test } from './tests'
+import { tests, Test } from './tests/tests'
+import { groupTests } from './tests/groupTests'
 
 type Result = 'waiting' | 'running' | 'success' | 'failure' | 'error'
 
@@ -97,22 +99,26 @@ function TestView({
 
 export default function TestScreen(): JSX.Element {
   const [completedTests, setCompletedTests] = useState<number>(0)
+  const route = useRoute();
+  const params = route.params as { onlyGroups: boolean };
+  const allTests = tests.concat(groupTests);
+  const activeTests = params.onlyGroups ? groupTests : allTests;
 
   return (
     <ScrollView>
       <View>
         <View style={{ padding: 12 }}>
           <Text testID="Test View" accessible accessibilityLabel="Test View">
-            Unit Tests
+          {params.onlyGroups ? 'Group Unit Tests' : 'All Unit Tests'}          
           </Text>
           <View
             style={{ flexDirection: 'row', justifyContent: 'space-between' }}
           >
             <Text>
-              Running {completedTests}/{tests.length}
+              Running {completedTests}/{activeTests.length}
             </Text>
 
-            {completedTests === tests.length && (
+            {completedTests === activeTests.length && (
               <Text
                 testID="tests-complete"
                 accessible
@@ -129,7 +135,7 @@ export default function TestScreen(): JSX.Element {
           accessibilityLabel="tests-complete"
           style={{ paddingHorizontal: 12 }}
         >
-          {(tests || []).slice(0, completedTests + 1).map((test: Test, i) => {
+          {(activeTests || []).slice(0, completedTests + 1).map((test: Test, i) => {
             return (
               <TestView
                 test={test}
